March 2023 Maintenance
Due to some problems in the TEMOA-Italy GitHub repository, it was canceled and substituted by a brand new one on April 2nd 2023.
Then, the repository was recreated together with the releases 1.0 and 2.0. Eventually, starting from release 2.0, this single commit was used to update the repository to the April 2nd 2023 version.
Starting from this update, .sqlite databases will be published within dedicated ZIP archives.