Update database_preprocessing.py
This commit adds to the database_preprocessing.py script a part looking for errors in the TechInputSplit, TechOutputSplit, MinInputGroup and MinOutputGroup values. Specifically, the script notifies to the modeler if the sum of such parameters for a specific region, period and technology/group is higher than 1 (since this would make impossible for the model to satisfy the associated constraint).